+++ channel/Channel.hpp
@@ -6,7 +6,7 @@
 /*   By: yli <yli@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 18:54:42 by yli               #+#    #+#             */
-/*   Updated: 2023/09/28 19:35:25 by yli              ###   ########.fr       */
+/*   Updated: 2023/10/03 00:26:59 by nmaliare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,9 +27,8 @@ class Channel
         Channel();
         ~Channel();
 
-
         void    setInviteOnly(bool invite_only);
-        void    setUserLimit(unsigned long long int size);
+        void    setUserLimit(int size);
         void    setChannelName(std::string& channelName);
         void    setKey(const std::string key);
         std::string getKey(void);
@@ -60,7 +59,7 @@ class Channel
 
     private:
         std::string _channelName;
-        unsigned long long int  _user_limit;
+        int    _user_limit;
         bool    _invite_only;
         bool    _topic_right;
         std::string _key;
